+++ src/amqp/amqp_consumer/multiple.rs
@@ -16,7 +16,7 @@ use crate::{
         amqp_client::AmqpClient,
         error::{DisposeConsumerError, RecoverAndReceiveError},
     },
-    core::{RecoverableError, RecoverableTransport, TransportClient},
+    core::{RecoverableError, RecoverableTransport},
     event_hubs_retry_policy::EventHubsRetryPolicy,
     util::{self, time::timeout},
     ReceivedEventData,
@@ -226,28 +226,28 @@ where
     }
 }
 
-async fn recover_consumers<RP>(
-    client: &mut AmqpClient,
-    consumers: &mut MultipleAmqpConsumers<RP>,
-) -> Result<(), RecoverAndReceiveError>
-where
-    RP: EventHubsRetryPolicy + Send,
-{
-    log::debug!("Recovering consumers");
+// async fn recover_consumers<RP>(
+//     client: &mut AmqpClient,
+//     consumers: &mut MultipleAmqpConsumers<RP>,
+// ) -> Result<(), RecoverAndReceiveError>
+// where
+//     RP: EventHubsRetryPolicy + Send,
+// {
+//     log::debug!("Recovering consumers");
 
-    // Client should be already recovered
-    let mut result = Ok(());
+//     // Client should be already recovered
+//     let mut result = Ok(());
 
-    for c in consumers.inner.iter_mut() {
-        // Consumers that are polling would probably encounter an error
-        // when they poll again, and it will enter another recovery
-        if let ConsumerState::Value { value } = c {
-            result = result.and(client.recover_consumer(value).await);
-        }
-    }
+//     for c in consumers.inner.iter_mut() {
+//         // Consumers that are polling would probably encounter an error
+//         // when they poll again, and it will enter another recovery
+//         if let ConsumerState::Value { value } = c {
+//             result = result.and(client.recover_consumer(value).await);
+//         }
+//     }
 
-    result.map_err(Into::into)
-}
+//     result.map_err(Into::into)
+// }
 
 async fn recover_consumers_by_creating_new_consumers<RP>(
     client: &mut AmqpClient,
@@ -276,7 +276,6 @@ async fn recover_and_recv<RP>(
     client: &mut AmqpClient,
     consumers: &mut MultipleAmqpConsumers<RP>,
     should_try_recover: bool,
-    should_resume_consumers: bool,
 ) -> Result<Option<ReceivedEventData>, RecoverAndReceiveError>
 where
     RP: EventHubsRetryPolicy + Send + Unpin + 'static,
@@ -289,10 +288,7 @@ where
             }
         }
 
-        match should_resume_consumers {
-            true => recover_consumers(client, consumers).await?,
-            false => recover_consumers_by_creating_new_consumers(client, consumers).await?,
-        }
+        recover_consumers_by_creating_new_consumers(client, consumers).await?;
     }
 
     consumers.recv().await.transpose().map_err(Into::into)
@@ -308,12 +304,11 @@ where
     let mut failed_attempts = 0;
     let mut try_timeout = consumers.retry_policy.calculate_try_timeout(failed_attempts);
     let mut should_try_recover = false;
-    let mut should_resume_consumers = true;
 
     loop {
         let err = match timeout(
             try_timeout,
-            recover_and_recv(client, consumers, should_try_recover, should_resume_consumers)
+            recover_and_recv(client, consumers, should_try_recover)
         ).await {
             Ok(result) => match result.transpose()? {
                 Ok(event) => return Some(Ok(event)),
@@ -332,7 +327,6 @@ where
             return Some(Err(err));
         }
         should_try_recover = err.should_try_recover();
-        should_resume_consumers = err.is_link_resumable();
 
         failed_attempts += 1;
         let retry_delay = consumers